Creating new classes and moving generate methods to respective classes
