+++ src/pynumad/objects/geometry.py
@@ -25,6 +25,8 @@ class Geometry:
         Chord distribution [m]
     chordoffset : array
         Chordwise offset (in addition to natural offset)
+    coordinates : array
+        actual x,y,z geometry
     degreestwist : array
         Twist distribution [degrees]
     ispan : array
@@ -43,9 +45,6 @@ class Geometry:
         Locations of keypoints b & c, defines distance
         between keypoints b & c [mm]. First entry is the HP spar cap.
         Second entry is the LP spar cap
-    stations : list
-        Blade Stations, define the camber and thickness along the blade,
-        refer to ``StationDef``
     sweep : array
         Blade Sweep, Reference axis location along x1 [m]
     teband : float
@@ -55,6 +54,8 @@ class Geometry:
         interpolated chord
     ipercentthick : array
         interpolated thickness
+    c : array
+        nondimensional (0-1) value of x axis
     ic : array
     icamber : array
     ithickness : array
@@ -70,41 +71,28 @@ class Geometry:
         natural offset
     profiles : array
         normalized airfoil profiles
-    geometry : array
-        actual x,y,z geometry
-    arclength : array
-        surface distance from L.E.
     cpos : array
-        chordwise position
-    hgGeometry : list
-    hgKeypoints : list
-
-    Example
-    -------
-    blade = BladeDef()
+        x-axis parametrization for airfoil curve
     """
 
-    def __init__(self, shape, settings=None):
-        n_points, n_stations = shape
-        self.c: ndarray = np.zeros((n_points, n_stations))
-        # nondimensional (0-1) value of x axis #
-        self.camber: ndarray = np.zeros((n_points, n_stations))
-        self.thickness: ndarray = np.zeros((n_points, n_stations))
-        self.ic: ndarray = np.zeros((n_points, n_stations))
-        self.icamber: ndarray = np.zeros((n_points, n_stations))
-        self.ithickness: ndarray = np.zeros((n_points, n_stations))
-        self.cpos = None
-        # cpos x-axis parametrization for airfoil curve
-        self.idegreestwist = None
-        self.ichord = None
+    def __init__(self, settings=None):
+        self.c: ndarray = None
+        self.camber: ndarray = None
+        self.thickness: ndarray = None
+        self.ic: ndarray = None
+        self.icamber: ndarray = None
+        self.ithickness: ndarray = None
+        self.cpos: ndarray = None
+        self.idegreestwist: ndarray = None
+        self.ichord: ndarray = None
         self.ichordoffset: ndarray = None
         self.iaerocenter: ndarray = None
         self.idegreestwist: ndarray = None
-        self.ipercentthick = None
-        self.profiles = None
-        self.coordinates = None
-        self.xoffset = None
-        self.LEindex = None
+        self.ipercentthick: ndarray = None
+        self.profiles: ndarray = None
+        self.coordinates: ndarray = None
+        self.xoffset: ndarray = None
+        self.LEindex: ndarray = None
         self.iprebend: ndarray = None
         self.isweep: ndarray = None
 
@@ -112,6 +100,236 @@ class Geometry:
         self._natural_offset: int = 1
         self._rotorspin: int = 1
         self._swtwisted: int = 0
+        
+    def initialize_arrays(self, num_points: int, num_stations: int, num_istations: int):
+        """_summary_
+
+        Parameters
+        ----------
+        size : tuple[int]
+            num_points, num_stations = size
+        """
+        size = (num_points, num_stations)
+        isize = (num_points, num_istations)        
+        self.c = np.zeros(size)
+        self.camber = np.zeros(size)
+        self.thickness = np.zeros(size)
+        self.ic = np.zeros(isize)
+        self.icamber = np.zeros(isize)
+        self.ithickness = np.zeros(isize)
+        self.idegreestwist = np.zeros(isize)
+        self.ichord = np.zeros(isize)
+        self.ichordoffset = np.zeros(isize)
+        self.iaerocenter = np.zeros(isize)
+        self.idegreestwist = np.zeros(isize)
+        self.ipercentthick = np.zeros(isize)
+        self.iprebend = np.zeros(isize)
+        self.isweep = np.zeros(isize)
+        
+        
+    def _compare(self, other):
+        """
+        Parameters
+        ----------
+        other : Geometry
+
+        Returns
+        -------
+        bool
+        """
+        attrs = [
+            a
+            for a in dir(self)
+            if not a.startswith("__") and not callable(getattr(self, a))
+        ]
+        for attr in attrs:
+            if getattr(self, attr) != getattr(other, attr):
+                return False
+        return True
+    
+    def generate(self, definition):
+        """Populates geometry attributes based on a given blade defintion
+
+        Parameters
+        ----------
+        definition : Definition
+
+        Returns
+        -------
+        self
+        """
+        self.ispan = definition.ispan
+        num_istations = self.ispan.size
+        stations = definition.stations
+        num_stations = len(stations)
+        if num_stations > 0:
+            num_points = len(stations[0].airfoil.c)
+        else:
+            raise Exception(
+                "BladeDef must have at least one station before updating geometry."
+            )
+
+        # first station must be at blade root to prevent extrapolation
+        assert stations[0].spanlocation == 0, "first station must be at the blade root"
+
+        self.initialize_arrays(num_points, num_stations, num_istations)
+
+        self._natural_offset = definition.natural_offset
+        self._rotorspin = definition.rotorspin
+        self._swtwisted = definition.swtwisted
+
+        # Collect parameter tables from the stations.
+        spanlocation = np.array(
+            [stations[i].spanlocation for i in range(len(stations))]
+        )
+        tetype = ["_"] * num_stations
+        for k in range(num_stations):
+            station = stations[k]
+            assert (
+                len(station.airfoil.c) == num_points
+            ), "Station airfoils must have same number of samples."
+
+            self.cf = station.airfoil.c
+            self.camber[:, k] = station.airfoil.camber
+            self.thickness[:, k] = station.airfoil.thickness
+            tetype[k] = station.airfoil.te_type
+
+        # fix numerical issue due to precision on camber calculation
+        # camber should start and end at y-values of zero
+        self.camber[0, :] = np.zeros((1, num_stations))
+        self.camber[-1, :] = np.zeros((1, num_stations))
+
+        # Interpolate the station parameter tables.
+        # Each column corresponds to an interpolated station.
+
+        ## ic
+        self.ic = interpolator_wrap(
+            spanlocation, self.c, self.ispan, "pchip", axis=1
+        )
+
+        ## cpos
+        self.cpos = np.concatenate(
+            (
+                -self.ic[-1, :].reshape(1, -1),
+                -np.flipud(self.ic),
+                self.ic[1:, :],
+                self.ic[-1, :].reshape(1, -1),
+            ),
+            axis=0,
+        )
+
+        ## icamber
+        self.icamber = interpolator_wrap(
+            spanlocation, self.camber, self.ispan, "pchip", axis=1
+        )
+
+        ## ithickness
+        self.ithickness = interpolator_wrap(
+            spanlocation, self.thickness, self.ispan, "pchip", axis=1
+        )
+        # Adjust the thickness profiles based on te_type of stations.
+        # This is mainly for transitions to flatbacks were the
+        # interpolated airfoil needs to look like a round.
+        for k in range(len(self.ispan)):
+            try:
+                ind = np.argwhere(self.ispan[k] < spanlocation)[0][0]
+                # maybe better: ind = np.flatnonzero(self.ispan[k] < spanlocation)[0]
+            except:
+                continue
+            else:
+                if ind == 1:
+                    continue
+            if tetype[ind] == "flat" and tetype[ind - 1] == "round":
+                self.ithickness[-1, k] = 0
+
+        # Interpolate the blade parameter curves.
+        ## idegreestwist
+        self.idegreestwist = interpolator_wrap(
+            definition.span, definition.degreestwist, self.ispan, "pchip"
+        )
+
+        ## ichord
+        self.ichord = interpolator_wrap(
+            definition.span, definition.chord, self.ispan, "pchip"
+        )
+
+        ## ipercentthick
+        absolutethick = np.multiply(definition.percentthick, definition.chord) / 100
+        iabsolutethick = interpolator_wrap(
+            definition.span, absolutethick, self.ispan, "pchip"
+        )
+        self.ipercentthick = iabsolutethick / self.ichord * 100
+        # ensure that the interpolation doesn't reduce the percent
+        # thickness beneath the thinnest airfoil
+        self.ipercentthick[
+            self.ipercentthick < np.amin(definition.percentthick)
+        ] = np.amin(definition.percentthick)
+
+        ## ichordoffset
+        self.ichordoffset = interpolator_wrap(
+            definition.span, definition.chordoffset, self.ispan, "pchip"
+        )
+
+        ## iaerocenter
+        self.iaerocenter = interpolator_wrap(
+            definition.span, definition.aerocenter, self.ispan, "pchip"
+        )
+
+        ## isweep
+        if len(definition.sweep) == 0:
+            definition.sweep = np.zeros((self.span.shape, self.span.shape))
+        if len(definition.prebend) == 0:
+            definition.prebend = np.zeros((self.span.shape, self.span.shape))
+
+        self.isweep = interpolator_wrap(
+            definition.span, definition.sweep, self.ispan, "pchip"
+        )
+
+        ## iprebend
+        self.iprebend = interpolator_wrap(
+            definition.span, definition.prebend, self.ispan, "pchip"
+        )
+
+        # Generate the blade surface self.
+        n_istations = np.asarray(self.ispan).size
+        n_areas = num_points * 2 + 1
+        self.profiles = np.zeros((n_areas, 2, n_istations))
+        self.coordinates = np.zeros((n_areas, 3, n_istations))
+        self.xoffset = np.zeros((1, n_istations))
+        self.LEindex = num_points
+
+        for k in range(n_istations):
+            self.update_airfoil_profile(k)
+            mtindex = np.argmax(self.ithickness[:, k])
+            self.xoffset[0, k] = self.ic[mtindex, k]
+            self.update_oml_geometry(k)
+
+        # Calculate the arc length of each curve
+        self.arclength = np.zeros((n_areas, n_istations))
+        self.HParcx0 = np.zeros((1, n_istations))
+        self.LParcx0 = np.zeros((1, n_istations))
+
+        le_index = self.LEindex
+        for k in range(n_istations):
+            xx = self.coordinates[:, 0, k]
+            yy = self.coordinates[:, 1, k]
+            zz = self.coordinates[:, 2, k]
+            arclen = np.sqrt(np.diff(xx) ** 2 + np.diff(yy) ** 2 + np.diff(zz) ** 2)
+            arclen = np.concatenate((np.array([0]), np.cumsum(arclen)), axis=0)
+            self.arclength[:, k] = arclen
+            LEarcsum = self.arclength[le_index, k]
+            self.arclength[:, k] = self.arclength[:, k] - LEarcsum
+
+            # find where x=0 intersects the surface
+            self.HParcx0[0, k] = (
+                interpolator_wrap(xx[1 : le_index + 1], arclen[1 : le_index + 1], 0) - LEarcsum
+            )
+            self.LParcx0[0, k] = (
+                interpolator_wrap(xx[-2 : le_index - 1 : -1], arclen[-2 : le_index - 1 : -1], 0)
+                - LEarcsum
+            )
+
+        return self
 
     def update_airfoil_profile(self, k):
         """_summary_
